refactoring the models and controllers for device status microservice
