+++ src/device-status/app.py
@@ -1,11 +1,14 @@
+from controllers.status import status_bp
+from controllers.health import health_bp
 from flask import Flask
 import logging
 import os
 from flask_cors import CORS
 from flask_pymongo import PyMongo
-from config import config, db_connection
+from config import constants, db_connection
 from dotenv import load_dotenv
 load_dotenv()
+# import blueprints
 
 
 _logger = logging.getLogger(__name__)
@@ -17,19 +20,15 @@ mongo = PyMongo()
 def create_app(environment):
     # create a flask app instance
     app = Flask(__name__)
-    app.config.from_object(config.app_config[environment])
+    app.config.from_object(constants.app_config[environment])
     mongo.init_app(app)
 
     # Allow cross-brower resource sharing
     CORS(app)
 
-    # import blueprints
-    from controllers.check_health import monitor_bp
-    from controllers.check_status import device_status_bp
-
     # register blueprints
-    app.register_blueprint(monitor_bp)
-    app.register_blueprint(device_status_bp)
+    app.register_blueprint(health_bp)
+    app.register_blueprint(status_bp)
 
     return app
